requirements.txt
torch>=2.0.1
torchvision>=0.15.2
lightning[extra]>=2.0.2
torchmetrics>=0.11.4
unidecode
inflect
matplotlib
scipy
pillow
einops
g2p-en
pyyaml
validators
onnx
onnxruntime
numpy
# needed for data preparation
librosa
unidecode
inflect
pydantic
pyworld
# for audio playback
sounddevice
# sudo apt-get install portaudio19-dev
typed-argument-parser
phonemizer

